// src/interfaces/persistent.ts
export interface ReadReq {
  options?: ReadReq_Options | undefined;
  ack?: ReadReq_Ack | undefined;
  nack?: ReadReq_Nack | undefined;
}

export interface ReadReq_Options {
  streamIdentifier?: StreamIdentifier | undefined;
  all?: Empty | undefined;
  groupName: string;
  bufferSize: number;
  uuidOption: ReadReq_Options_UUIDOption | undefined;
}

export interface ReadReq_Options_UUIDOption {
  structured?: Empty | undefined;
  string?: Empty | undefined;
}

export interface ReadReq_Ack {
  id: Uint8Array;
  ids: UUID[];
}

export interface ReadReq_Nack {
  id: Uint8Array;
  ids: UUID[];
  action: ReadReq_Nack_Action;
  reason: string;
}

export enum ReadReq_Nack_Action {
  Unknown = 0,
  Park = 1,
  Retry = 2,
  Skip = 3,
  Stop = 4,
  UNRECOGNIZED = -1,
}

export interface ReadResp {
  event?: ReadResp_ReadEvent | undefined;
  subscriptionConfirmation?: ReadResp_SubscriptionConfirmation | undefined;
}

export interface ReadResp_ReadEvent {
  event: ReadResp_ReadEvent_RecordedEvent | undefined;
  link: ReadResp_ReadEvent_RecordedEvent | undefined;
  commitPosition?: number | undefined;
  noPosition?: Empty | undefined;
  retryCount?: number | undefined;
  noRetryCount?: Empty | undefined;
}

export interface ReadResp_ReadEvent_RecordedEvent {
  id: UUID | undefined;
  streamIdentifier: StreamIdentifier | undefined;
  streamRevision: number;
  preparePosition: number;
  commitPosition: number;
  metadata: { [key: string]: string };
  customMetadata: Uint8Array;
  data: Uint8Array;
}

export interface ReadResp_ReadEvent_RecordedEvent_MetadataEntry {
  key: string;
  value: string;
}

export interface ReadResp_SubscriptionConfirmation {
  subscriptionId: string;
}

export interface CreateReq {
  options: CreateReq_Options | undefined;
}

export enum CreateReq_ConsumerStrategy {
  DispatchToSingle = 0,
  RoundRobin = 1,
  Pinned = 2,
  UNRECOGNIZED = -1,
}

export interface CreateReq_Options {
  stream?: CreateReq_StreamOptions | undefined;
  all?: CreateReq_AllOptions | undefined;
  /** @deprecated */
  streamIdentifier: StreamIdentifier | undefined;
  groupName: string;
  settings: CreateReq_Settings | undefined;
}

export interface CreateReq_StreamOptions {
  streamIdentifier: StreamIdentifier | undefined;
  revision?: number | undefined;
  start?: Empty | undefined;
  end?: Empty | undefined;
}

export interface CreateReq_AllOptions {
  position?: CreateReq_Position | undefined;
  start?: Empty | undefined;
  end?: Empty | undefined;
  filter?: CreateReq_AllOptions_FilterOptions | undefined;
  noFilter?: Empty | undefined;
}
// ...
export interface CreateReq_Settings {
  resolveLinks: boolean;
  /** @deprecated */
  revision: number;
  extraStatistics: boolean;
  maxRetryCount: number;
  minCheckpointCount: number;
  maxCheckpointCount: number;
  maxSubscriberCount: number;
  liveBufferSize: number;
  readBatchSize: number;
  historyBufferSize: number;
  /** @deprecated */
  namedConsumerStrategy: CreateReq_ConsumerStrategy;
  messageTimeoutTicks?: number | undefined;
  messageTimeoutMs?: number | undefined;
  checkpointAfterTicks?: number | undefined;
  checkpointAfterMs?: number | undefined;
  consumerStrategy: string;
}

export interface CreateResp {}

export interface UpdateReq {
  options: UpdateReq_Options | undefined;
}

export enum UpdateReq_ConsumerStrategy {
  DispatchToSingle = 0,
  RoundRobin = 1,
  Pinned = 2,
  UNRECOGNIZED = -1,
}

export interface UpdateReq_Options {
  stream?: UpdateReq_StreamOptions | undefined;
  all?: UpdateReq_AllOptions | undefined;
  /** @deprecated */
  streamIdentifier: StreamIdentifier | undefined;
  groupName: string;
  settings: UpdateReq_Settings | undefined;
}

export interface UpdateReq_StreamOptions {
  streamIdentifier: StreamIdentifier | undefined;
  revision?: number | undefined;
  start?: Empty | undefined;
  end?: Empty | undefined;
}

export interface UpdateReq_AllOptions {
  position?: UpdateReq_Position | undefined;
  start?: Empty | undefined;
  end?: Empty | undefined;
}

export interface UpdateReq_Position {
  commitPosition: number;
  preparePosition: number;
}

export interface UpdateReq_Settings {
  resolveLinks: boolean;
  /** @deprecated */
  revision: number;
  extraStatistics: boolean;
  maxRetryCount: number;
  minCheckpointCount: number;
  maxCheckpointCount: number;
  maxSubscriberCount: number;
  liveBufferSize: number;
  readBatchSize: number;
  historyBufferSize: number;
  namedConsumerStrategy: UpdateReq_ConsumerStrategy;
  messageTimeoutTicks?: number | undefined;
  messageTimeoutMs?: number | undefined;
  checkpointAfterTicks?: number | undefined;
  checkpointAfterMs?: number | undefined;
}

export interface UpdateResp {}

export interface DeleteReq {
  options: DeleteReq_Options | undefined;
}

export interface DeleteReq_Options {
  streamIdentifier?: StreamIdentifier | undefined;
  all?: Empty | undefined;
  groupName: string;
}

export interface DeleteResp {}

export interface GetInfoReq {
  options: GetInfoReq_Options | undefined;
}

export interface GetInfoReq_Options {
  streamIdentifier?: StreamIdentifier | undefined;
  all?: Empty | undefined;
  groupName: string;
}

export interface GetInfoResp {
  subscriptionInfo: SubscriptionInfo | undefined;
}

export interface SubscriptionInfo {
  eventSource: string;
  groupName: string;
  status: string;
  connections: SubscriptionInfo_ConnectionInfo[];
  averagePerSecond: number;
  totalItems: number;
  countSinceLastMeasurement: number;
  lastCheckpointedEventPosition: string;
  lastKnownEventPosition: string;
  resolveLinkTos: boolean;
  startFrom: string;
  messageTimeoutMilliseconds: number;
  extraStatistics: boolean;
  maxRetryCount: number;
  liveBufferSize: number;
  bufferSize: number;
  readBatchSize: number;
  checkPointAfterMilliseconds: number;
  minCheckPointCount: number;
  maxCheckPointCount: number;
  readBufferCount: number;
  liveBufferCount: number;
  retryBufferCount: number;
  totalInFlightMessages: number;
  outstandingMessagesCount: number;
  namedConsumerStrategy: string;
  maxSubscriberCount: number;
  parkedMessageCount: number;
}

export interface SubscriptionInfo_ConnectionInfo {
  from: string;
  username: string;
  averageItemsPerSecond: number;
  totalItems: number;
  countSinceLastMeasurement: number;
  observedMeasurements: SubscriptionInfo_Measurement[];
  availableSlots: number;
  inFlightMessages: number;
  connectionName: string;
}

export interface SubscriptionInfo_Measurement {
  key: string;
  value: number;
}

export interface ReplayParkedReq {
  options: ReplayParkedReq_Options | undefined;
}

export interface ReplayParkedReq_Options {
  groupName: string;
  streamIdentifier?: StreamIdentifier | undefined;
  all?: Empty | undefined;
  stopAt?: number | undefined;
  noLimit?: Empty | undefined;
}

export interface ReplayParkedResp {}

export interface ListReq {
  options: ListReq_Options | undefined;
}

export interface ListReq_Options {
  listAllSubscriptions?: Empty | undefined;
  listForStream?: ListReq_StreamOption | undefined;
}

export interface ListReq_StreamOption {
  stream?: StreamIdentifier | undefined;
  all?: Empty | undefined;
}

export interface ListResp {
  subscriptions: SubscriptionInfo[];
}
// ...
export interface PersistentSubscriptionsController {
  create(
    request: CreateReq,
  ): Promise<CreateResp> | Observable<CreateResp> | CreateResp;

  update(
    request: UpdateReq,
  ): Promise<UpdateResp> | Observable<UpdateResp> | UpdateResp;

  delete(
    request: DeleteReq,
  ): Promise<DeleteResp> | Observable<DeleteResp> | DeleteResp;

  read(request: Observable<ReadReq>): Observable<ReadResp>;

  getInfo(
    request: GetInfoReq,
  ): Promise<GetInfoResp> | Observable<GetInfoResp> | GetInfoResp;

  replayParked(
    request: ReplayParkedReq,
  ):
    | Promise<ReplayParkedResp>
    | Observable<ReplayParkedResp>
    | ReplayParkedResp;

  list(request: ListReq): Promise<ListResp> | Observable<ListResp> | ListResp;

  restartSubsystem(request: Empty): Promise<Empty> | Observable<Empty> | Empty;
}